requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
email-validator==2.2.0